requirements.txt
# ── Web Framework ─────────────────────────────────────────────────────────────
fastapi==0.115.6
uvicorn[standard]==0.32.1
python-multipart==0.0.20         # file upload support

# ── Video / Audio I/O ─────────────────────────────────────────────────────────
yt-dlp==2024.12.13
ffmpeg-python==0.2.0

# ── Speech: Transcription + Diarization ───────────────────────────────────────
openai-whisper==20240930
pyannote.audio==3.3.2
torch==2.5.1
torchaudio==2.5.1

# ── Vision: Frame Comparison + VLM ────────────────────────────────────────────
opencv-python-headless==4.10.0.84
Pillow==11.1.0
open-clip-torch==2.29.0          # CLIP (openai/clip-vit-base-patch32)
transformers==4.47.1             # Qwen2.5-VL + Mistral
accelerate==1.2.1                # multi-GPU / device_map="auto"
bitsandbytes==0.45.0             # 4-bit / 8-bit quantisation
sentencepiece==0.2.0             # Mistral tokeniser
qwen-vl-utils==0.0.10            # Qwen2.5-VL image pre-processing helper

# ── PDF Output ────────────────────────────────────────────────────────────────
weasyprint==63.1
Jinja2==3.1.5                    # HTML → PDF template engine

# ── Storage ───────────────────────────────────────────────────────────────────
# sqlite3 is part of the Python standard library — no extra install needed

# ── Utilities ─────────────────────────────────────────────────────────────────
numpy==2.1.3
tqdm==4.67.1
python-dotenv==1.0.1             # .env support for secrets (HF tokens etc.)
aiofiles==24.1.0                 # async file I/O for FastAPI
httpx==0.28.1                    # async HTTP client (used in tests / health checks)

# ── DevOps (droplet_manager.py) ───────────────────────────────────────────────
paramiko==3.5.0                  # SSH / SFTP
requests==2.32.3                 # Digital Ocean API v2 çağrıları
